# openapi.yaml
...
openapi: 3.0.0
info:
  x-skaffolder-id-project: 5e5d475052a8e9561a0d35dc
  title: School
  description: You can obtain JWT token from login API
  version: 0.1
servers:
  - url: http://localhost:3000/api
    description: Local server for testing
paths:
...
  "/course/{id}":
    delete:
      x-skaffolder-id: 5e5e9a70526d3b6c7db6863f
      x-skaffolder-name: delete
      x-skaffolder-id-resource: 5e5df994526d3b6c7db67ebf
      x-skaffolder-resource: Course
      x-skaffolder-crudAction: delete
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION delete
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION delete
      tags:
        - Course
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    get:
      x-skaffolder-id: 5e5e9a6f526d3b6c7db68636
      x-skaffolder-name: get
      x-skaffolder-id-resource: 5e5df994526d3b6c7db67ebf
      x-skaffolder-resource: Course
      x-skaffolder-crudAction: get
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION get
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION get
      tags:
        - Course
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id resource
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Course"
    post:
      x-skaffolder-id: 5e5e9a6f526d3b6c7db68639
      x-skaffolder-name: update
      x-skaffolder-id-resource: 5e5df994526d3b6c7db67ebf
      x-skaffolder-resource: Course
      x-skaffolder-crudAction: update
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION update
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION update
      tags:
        - Course
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Course"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Course"
...
  "/events/{id}":
    delete:
      x-skaffolder-id: 5e5df0af526d3b6c7db67e8d
      x-skaffolder-name: delete
      x-skaffolder-id-resource: 5e5df0a1526d3b6c7db67e81
      x-skaffolder-resource: Events
      x-skaffolder-crudAction: delete
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION delete
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION delete
      tags:
        - Events
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    get:
      x-skaffolder-id: 5e5df0af526d3b6c7db67e87
      x-skaffolder-name: get
      x-skaffolder-id-resource: 5e5df0a1526d3b6c7db67e81
      x-skaffolder-resource: Events
      x-skaffolder-crudAction: get
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION get
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION get
      tags:
        - Events
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id resource
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Events"
    post:
      x-skaffolder-id: 5e5df0af526d3b6c7db67e89
      x-skaffolder-name: update
      x-skaffolder-id-resource: 5e5df0a1526d3b6c7db67e81
      x-skaffolder-resource: Events
      x-skaffolder-crudAction: update
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION update
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION update
      tags:
        - Events
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Events"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Events"
  "/events/findBy_student/{key}":
    get:
      x-skaffolder-id: 5e5e9a65526d3b6c7db68633
      x-skaffolder-name: findBy_student
      x-skaffolder-id-resource: 5e5df0a1526d3b6c7db67e81
      x-skaffolder-resource: Events
      x-skaffolder-crudAction: findBy_student
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION findBy_student
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /findBy_student/{key}
      x-skaffolder-roles: []
      summary: CRUD ACTION findBy_student
      tags:
        - Events
      parameters:
        - name: key
          x-skaffolder-type: Objectid
          in: path
          description: Id of model to search for
          required: true
          schema:
            type: Objectid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Events"
...
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Course:
      x-skaffolder-id: 5e5df994526d3b6c7db67ebf
      x-skaffolder-id-db: 5e5d475052a8e9561a0d35dd
      x-skaffolder-id-entity: 5e5df994526d3b6c7db67ebd
      x-skaffolder-url: /course
      properties:
        _id:
          type: string
          x-skaffolder-required: true
        name:
          type: string
          x-skaffolder-id-attr: 5e5df994526d3b6c7db67ebe
          x-skaffolder-type: String
          x-skaffolder-required: true
      x-skaffolder-relations: 
      required:
        - _id
        - name
    Events:
      x-skaffolder-id: 5e5df0a1526d3b6c7db67e81
      x-skaffolder-id-db: 5e5d475052a8e9561a0d35dd
      x-skaffolder-id-entity: 5e5df0a1526d3b6c7db67e7d
      x-skaffolder-url: /events
      properties:
        _id:
          type: string
          x-skaffolder-required: true
        date:
          type: integer
          x-skaffolder-id-attr: 5e5df0a1526d3b6c7db67e7f
          x-skaffolder-type: Date
        description:
          type: string
          x-skaffolder-id-attr: 5e5df0a1526d3b6c7db67e80
          x-skaffolder-type: String
        name:
          type: string
          x-skaffolder-id-attr: 5e5df0a1526d3b6c7db67e7e
          x-skaffolder-type: String
          x-skaffolder-required: true
      x-skaffolder-relations:
        _student:
          x-skaffolder-id: 5e5e9a56526d3b6c7db68630
          x-skaffolder-type: m:m
          x-skaffolder-ent1: 5e5df0a1526d3b6c7db67e7d
          x-skaffolder-ent2: 5e5d48ca52a8e9561a0d3652
        _user:
          x-skaffolder-id: 5e5e9a56526d3b6c7db6862f
          x-skaffolder-type: m:m
          x-skaffolder-ent1: 5e5df0a1526d3b6c7db67e7d
          x-skaffolder-ent2: 5e5d475052a8e9561a0d35de
      required:
        - _id
        - name
...